LPS-89672 Move openImageSelector to FragmentsEditorDialogUtils
